+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- official test cases for good data at https://github.com/maxmind/MaxMind-DB/
+### Changed
+- string and binary IP address parsing to handle ranges and shortened addresses
+
 ## [1.3.1] - 2018-05-04
 ### Fixed
 - incompatibility with OTP 20.3.1 due to gen_statem bug (issue with init actions)